# server/docs/swagger.yaml
basePath: /api/
definitions:
...
  model.InstallConfig:
    properties:
      isInitDb:
        type: boolean
      mysql:
        $ref: '#/definitions/model.Mysql'
      redis:
        properties:
          addr:
            description: 服务器地址:端口
            type: string
          db:
            description: redis的哪个数据库
            type: integer
          password:
            description: 密码
            type: string
        type: object
    type: object
  model.Mysql:
    properties:
      config:
        description: 高级配置
        type: string
      dbName:
        description: 数据库名
        type: string
      logMode:
        description: 是否开启Gorm全局日志
        type: string
      logZap:
        description: 是否通过zap写入日志文件
        type: boolean
      maxIdleConns:
        description: 空闲中的最大连接数
        type: integer
      maxOpenConns:
        description: 打开到数据库的最大连接数
        type: integer
      password:
        description: 数据库密码
        type: string
      path:
        description: 服务器地址:端口
        type: string
      username:
        description: 数据库用户名
        type: string
    type: object
...
  request.CasbinInReceive:
    properties:
      casbinInfos:
        items:
          $ref: '#/definitions/request.CasbinInfo'
        type: array
      roleId:
        description: 权限id
        type: integer
    type: object
  request.CasbinInfo:
    properties:
      method:
        description: 方法
        type: string
      path:
        description: 路径
        type: string
    type: object
  request.Empty:
    type: object
...
  request.Login:
    properties:
      captchaId:
        description: 验证码ID
        type: string
      code:
        description: 验证码
        type: string
      mobile:
        description: 手机号
        type: string
      password:
        description: 密码
        type: string
      smsCode:
        description: 手机验证码
        type: string
      username:
        description: 用户名
        type: string
    type: object
...
  request.Register:
    properties:
      avatar:
        type: string
      email:
        type: string
      mobile:
        type: string
      nickname:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - mobile
    - password
    - username
    type: object
...
  request.UserChangePasswordStruct:
    properties:
      newPassword:
        description: 新密码
        type: string
      password:
        description: 密码
        type: string
    type: object
...
  request.UserUpdateAccountRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: integer
      mobile:
        type: string
      nickname:
        type: string
      username:
        type: string
    required:
    - avatar
    - email
    - mobile
    - nickname
    - username
    type: object
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /api/v1/common/auth/getUserMenus:
    post:
      parameters:
      - description: 空
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Empty'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取用户动态路由
      tags:
      - Auth
...
  /install/getServerInfo:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取服务器信息
      tags:
      - Install
  /install/getSystemConfig:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取配置文件表单
      tags:
      - Install
...
  /install/pong:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"安装成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: pong
      tags:
      - Install
  /install/setSystemConfig:
    post:
      parameters:
      - description: 设置配置文件内容
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.InstallConfig'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"设置成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 设置配置文件内容
      tags:
      - Install
...
  /v1/common/auth/changePassword:
    post:
      parameters:
      - description: 用户名, 原密码, 新密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UserChangePasswordStruct'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"修改成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用户修改密码
      tags:
      - Auth
  /v1/common/auth/getUserInfo:
    post:
      consumes:
      - application/json
      parameters:
      - description: 空
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Empty'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - Auth
  /v1/common/auth/login:
    post:
      parameters:
      - description: 用户名, 密码, 验证码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"登陆成功"}'
          schema:
            type: string
      summary: 用户登录
      tags:
      - Auth
  /v1/common/auth/logout:
    post:
      parameters:
      - description: 空
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Empty'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 退出登录
      tags:
      - Auth
  /v1/common/auth/register:
    post:
      parameters:
      - description: 用户名, 昵称, 密码, 角色ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Register'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"注册成功"}'
          schema:
            type: string
      summary: 用户注册账号
      tags:
      - Auth
  /v1/common/auth/update:
    post:
      parameters:
      - description: 用户名, 邮箱, 昵称，头像
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UserUpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"修改成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用户修改资料
      tags:
      - Auth
...
  /v1/system/casbin/UpdateCasbin:
    post:
      consumes:
      - application/json
      parameters:
      - description: 权限id, 权限模型列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CasbinInReceive'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新角色api权限
      tags:
      - Casbin
...
  /v1/system/getPolicyPathByRoleId:
    post:
      consumes:
      - application/json
      parameters:
      - description: 权限id, 权限模型列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CasbinInReceive'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取权限列表
      tags:
      - Casbin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"