# server/modules/install/model/sys_install.go
package model

type InstallConfig struct {
	Redis struct {
		DB       int    `mapstructure:"db" json:"db" yaml:"db"`                   // redis的哪个数据库
		Addr     string `mapstructure:"addr" json:"addr" yaml:"addr"`             // 服务器地址:端口
		Password string `mapstructure:"password" json:"password" yaml:"password"` // 密码
	}
	Mysql  Mysql
	IsInitDb bool `mapstructure:"isInitDb" json:"isInitDb"`
}

type Mysql struct {
	Path         string `mapstructure:"path" json:"path" yaml:"path"`                             // 服务器地址:端口
	Config       string `mapstructure:"config" json:"config" yaml:"config"`                       // 高级配置
	Dbname       string `mapstructure:"db-name" json:"dbName" yaml:"db-name"`                     // 数据库名
	Username     string `mapstructure:"username" json:"username" yaml:"username"`                 // 数据库用户名
	Password     string `mapstructure:"password" json:"password" yaml:"password"`                 // 数据库密码
	MaxIdleConns int    `mapstructure:"max-idle-conns" json:"maxIdleConns" yaml:"max-idle-conns"` // 空闲中的最大连接数
	MaxOpenConns int    `mapstructure:"max-open-conns" json:"maxOpenConns" yaml:"max-open-conns"` // 打开到数据库的最大连接数
	LogMode      string `mapstructure:"log-mode" json:"logMode" yaml:"log-mode"`                  // 是否开启Gorm全局日志
	LogZap       bool   `mapstructure:"log-zap" json:"logZap" yaml:"log-zap"`                     // 是否通过zap写入日志文件
}